Crate scad_syntax

A parser and lexer for the OpenSCAD language.

There are 4 key components in this crate,

SyntaxKind: an enum representing the various terminal and non-terminal items in the OpenSCAD grammar

the ast module: strongly-typed wrappers around the loosely-typed rowan::SyntaxNode

tokenize() - Convert a string into a stream of tokens (a SyntaxKind and its text)

parse() - Parse a stream of tokens into the top-level ast::Package it contains

Here is a typical use of this crate:

let src = r#"
  use <MCAD/math.scad>

  x = 1;
  y = 2;
  x_plus_y = x + y;
"#;

let tokens = scad_syntax::tokenize(src);
let (package, errors) = scad_syntax::parse(tokens);

assert!(errors.is_empty(), "{errors:?}");

An important thing to notice is that we will always get a ast::Package back from parsing, even if the tokens contain utter garbage. The parser borrows a lot of inspiration from rust-analyzer, so a lot of things mentioned in their architecture docs will also apply here.

In particular, we have the following design goals:

Syntax trees are lossless, or full fidelity. All comments and whitespace get preserved.

Syntax trees are semantic-less. They describe strictly the structure of a sequence of characters, they don’t have hygiene, name resolution or type information attached.

Syntax trees are simple value types. It is possible to create trees for a syntax without any external context.

Syntax trees have intuitive traversal API (parent, children, siblings, etc).

Parsing is lossless (even if the input is invalid, the tree produced by the parser represents it exactly).

Parsing is resilient (even if the input is invalid, parser tries to see as much syntax tree fragments in the input as it can).

Modules

ast

Automatically generated, strongly-typed AstNodes.

Macros

T

A helper macro for getting the SyntaxKind that corresponds to a particular token.

Structs

OpenSCAD

A tag type used by rowan to represent the OpenSCAD language.

ParseError

A basic parser error.

Enums

SyntaxKind

The different types of terminals and non-terminals in the OpenSCAD language grammar.

Functions

parse

Parse a set of tokens into a Package.

tokenize

Split the input text into its tokens.

Type Definitions

SyntaxNode

SyntaxToken